requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
sqlalchemy==1.4.11
python-dotenv==0.17.0
pymysql==1.0.2
cryptography==3.4.7
PyJWT==1.7.1
passlib==1.7.2